citas

+++ resources/views/home.blade.php
@@ -487,15 +487,15 @@ Permite ver todos los reportes al administrador
                     <button type="button" class="btn btn-dark" data-dismiss="modal">Cancelar</button>
                     <button id="confirm-delete-date" type="button" class="btn btn-danger" onclick="confirmDeleteDateButon()" >
                     
-                  
+                  @if($citas != [])
                         <a onclick="confirmDeleteDateButon()" class="text-decoration-none text-light" href="/deleteDate/{{ $cita->id }}">
                             Confirmar
                         </a>
-             
+                    @else
                     <a onclick="confirmDeleteDateButon()" class="text-decoration-none text-light" href="/deleteDate/">
                             Confirmar
                         </a>
-
+                    @endif
                     
                     </button>
